Add request ID to entity generation
Add a request_id field to the GenerateEntity and EntityGenerated
messages. When Scrimmage generates an entity at runtime, it will copy
the request ID from the GenerateEntity message to the EntityGenerated
message. This is to allow plugins that generate entities to be able to
associate a GenerateEntity request with the entity that was generated
in response